feat: iniciar validação na Marca

+++ app/Http/Controllers/BrandController.php
@@ -7,32 +7,57 @@ use Illuminate\Http\Request;
 
 class BrandController extends Controller
 {
+    private Brand $brand;
+
+    public function __construct(Brand $brand)
+    {
+        $this->brand = $brand;
+    }
+
     public function index()
     {
-        $brands = Brand::all();
-        return $brands;
+        // $brands = Brand::all();
+        $brands = $this->brand->all();
+        return response()->json($brands, 200);
     }
 
     public function store(Request $request)
     {
-        $brands = Brand::create($request->all());
-        return $brands;
+        // $brands = Brand::create($request->all());
+        $brand = $this->brand->create($request->all());
+        return response()->json($brand, 201);
     }
 
-    public function show(Brand $brand)
+    public function show($id)
     {
-        return $brand;
+        // $brand = Brand::find($id);
+        $brand = $this->brand->find($id);
+        if ($brand === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe!'], 404);
+        }
+
+        return response()->json($brand, 200);
     }
 
-    public function update(Request $request, Brand $brand)
+    public function update(Request $request, $id)
     {
+        $brand = $this->brand->find($id);
+
+        if ($brand === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe!'], 404);
+        }
+
         $brand->update($request->all());
-        return $brand;
+        return response()->json($brand, 200);
     }
 
-    public function destroy(Brand $brand)
+    public function destroy($id)
     {
+        $brand = $this->brand->find($id);
+        if ($brand === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
         $brand->delete();
-        return $brand;
+        return response()->json(["msg" => "A marca foi removida com sucesso!"], 200);
     }
 }
